+++ storage/remote/otlptranslator/prometheus/normalize_name.go
@@ -22,7 +22,6 @@ import (
 	"strings"
 	"unicode"
 
-	"github.com/prometheus/prometheus/util/strutil"
 	"go.opentelemetry.io/collector/pdata/pmetric"
 )
 
@@ -120,18 +119,19 @@ func BuildCompliantName(metric pmetric.Metric, namespace string, addMetricSuffix
 	return metricName
 }
 
+var nonMetricNameCharRE = regexp.MustCompile(`[^a-zA-Z0-9:]`)
+
 // Build a normalized name for the specified metric.
 func normalizeName(metric pmetric.Metric, namespace string, allowUTF8 bool) string {
 	var nameTokens []string
 	var separators []string
 	if !allowUTF8 {
-		nonTokenMetricCharRE := regexp.MustCompile(`[^a-zA-Z0-9:]`)
 		// Split metric name into "tokens" (of supported metric name runes).
 		// Note that this has the side effect of replacing multiple consecutive underscores with a single underscore.
 		// This is part of the OTel to Prometheus specification: https://github.com/open-telemetry/opentelemetry-specification/blob/v1.38.0/specification/compatibility/prometheus_and_openmetrics.md#otlp-metric-points-to-prometheus.
 		nameTokens = strings.FieldsFunc(
 			metric.Name(),
-			func(r rune) bool { return nonTokenMetricCharRE.MatchString(string(r)) },
+			func(r rune) bool { return nonMetricNameCharRE.MatchString(string(r)) },
 		)
 	} else {
 		translationFunc := func(r rune) bool { return !unicode.IsLetter(r) && !unicode.IsDigit(r) && r != ':' }
@@ -289,7 +289,7 @@ func cleanUpUnit(unit string) string {
 	// This is part of the OTel to Prometheus specification: https://github.com/open-telemetry/opentelemetry-specification/blob/v1.38.0/specification/compatibility/prometheus_and_openmetrics.md#otlp-metric-points-to-prometheus.
 	multipleUnderscoresRE := regexp.MustCompile(`__+`)
 	return strings.TrimPrefix(multipleUnderscoresRE.ReplaceAllString(
-		strutil.SanitizeLabelName(unit),
+		nonMetricNameCharRE.ReplaceAllString(unit, "_"),
 		"_",
 	), "_")
 }